Fix calls param not being parsed
#581
